chore: migrate more to TS
