+++ packages/cli-testing-library/src/index.js
@@ -28,10 +28,10 @@ export * from './config'
 export * from './helpers.js'
 export * from './events.js'
 export * from './get-queries-for-instance.js'
-export * from './matches'
+export * from './matches.js'
 export * from './pure'
 export * from './query-helpers'
 export * from './queries/index.js'
-export * from './mutation-observer'
+export * from './mutation-observer.js'
 export * from './wait-for.js'
 export * from './user-event'
