+++ packages/cli-testing-library/src/matchers/to-be-in-the-console.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-invalid-this */
-import {getDefaultNormalizer} from '../matches'
+import {getDefaultNormalizer} from '../matches.js'
 import {checkCliInstance, getMessage} from './utils'
 
 export function toBeInTheConsole(instance) {
