+++ packages/cli-testing-library/src/suggestions.ts
@@ -2,15 +2,15 @@ import {getDefaultNormalizer} from './matches'
 
 const normalize = getDefaultNormalizer()
 
-function escapeRegExp(string) {
+function escapeRegExp(string: string) {
   return string.replace(/[.*+\-?^${}()|[\]\\]/g, '\\$&') // $& means the whole matched string
 }
 
-function getRegExpMatcher(string) {
+function getRegExpMatcher(string: string) {
   return new RegExp(escapeRegExp(string.toLowerCase()), 'i')
 }
 
-function makeSuggestion(queryName, element, content, {variant, name}) {
+function makeSuggestion(queryName: string, element, content, {variant, name}) {
   const warning = ''
   const queryOptions = {}
   const queryArgs = [
@@ -52,7 +52,7 @@ function makeSuggestion(queryName, element, content, {variant, name}) {
   }
 }
 
-function canSuggest(currentMethod, requestedMethod, data) {
+function canSuggest(currentMethod: string, requestedMethod: string, data) {
   return (
     data &&
     (!requestedMethod ||
